fix: add UpDownCounterMetric.ts
